+++ sphinx-docs/changelog.md
@@ -12,6 +12,13 @@ Hlink adheres to semantic versioning as much as possible.
   find the documentation for this option in the new [Household Matching][household-matching-docs]
   section on the Configuration page.
 
+### Fixed
+
+* Fixed a bug in the calculation of predicted matches. Previously, if there was
+a second-best probability, hlink applied the threshold ratio only if
+the first and second-best probabilities were both at least at the alpha threshold. Now it always
+applies the threshold ratio when the best probability is at least at
+the alpha threshold and there is a second-best probability. [PR #200][pr200]
 
 ## v4.0.0 (2025-04-07)
 
@@ -380,6 +387,7 @@ and false negative data in model exploration. [PR #1][pr1]
 [pr184]: https://github.com/ipums/hlink/pull/184
 [pr185]: https://github.com/ipums/hlink/pull/185
 [pr189]: https://github.com/ipums/hlink/pull/189
+[pr200]: https://github.com/ipums/hlink/pull/200
 
 [household-matching-docs]: config.html#household-matching
 [ints-to-longs-docs]: config.html#data-sources